Using cookies for login, but so far the only way I could figure it out. The PUT and GET commands all work

+++ Data/DataAccess.cs
@@ -7,42 +7,53 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Formatters.Internal;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using HeyDo.Controllers;
 
 namespace HeyDo.Data
 {
     public class DataAccess
     {
         private static HttpClient _client = new HttpClient();
-        private static string baseUrl = "";
+        private static string baseUrl = "/";
+        
         
         //TODO clean up and test
-        public static async Task<JObject> ApiGoogle(string method, string json, string sub, string auth)
+        public static async Task<JObject> ApiGoogle(string method, string json, string sub, Dictionary<string,string> auth)
         {
-            var url = baseUrl + sub + ".json?";
-            var res = new HttpResponseMessage();
-            switch (method)
+            //Make sure user is authorized
+            if (await AuthController.Google(auth["token"]) == auth["uid"])
             {
-                case "PUT":
-                    HttpContent newContent = new StringContent(json, Encoding.UTF8, "application/json");
-                    res = await _client.PutAsync(url, newContent);
-                    break;
-                case "GET":
-                    res = await _client.GetAsync(url);
-                    break;
-                case "PATCH":
-                    HttpContent updateContent = new StringContent(json, Encoding.UTF8, "application/json");
-                    res = await _client.PatchAsync(url, updateContent);
-                    break;
-                case "DELETE":
-                    res = await _client.DeleteAsync(url);
-                    break;
-                default:
+                var url = baseUrl + sub + ".json?auth="+auth["token"];
+                var res = new HttpResponseMessage();
+                switch (method)
+                {
+                    case "PUT":
+                        HttpContent newContent = new StringContent(json, Encoding.UTF8, "application/json");
+                        res = await _client.PutAsync(url, newContent);
                         break;
+                    case "GET":
+                        res = await _client.GetAsync(url);
+                        break;
+                    case "PATCH":
+                        HttpContent updateContent = new StringContent(json, Encoding.UTF8, "application/json");
+                        res = await _client.PatchAsync(url, updateContent);
+                        break;
+                    case "DELETE":
+                        res = await _client.DeleteAsync(url);
+                        break;
+                    default:
+                            break;
+                }
+
+                var interim = await res.Content.ReadAsStringAsync();
+
+                return JsonConvert.DeserializeObject<JObject>(interim);
+            }
+            else
+            {
+                return new JObject();
             }
 
-            var interim = await res.Content.ReadAsStringAsync();
-
-            return JsonConvert.DeserializeObject<JObject>(interim);
         }
 
     }
